Split lens instances into a separate package
