+++ cabal.project
@@ -1,0 +1,3 @@
+packages:
+  monoidal-containers/
+  monoidal-containers-lens/
