+++ monoidal-containers/monoidal-containers.cabal
@@ -52,7 +52,6 @@ library
                        deepseq >=1.3 && <1.5,
                        unordered-containers >= 0.2 && < 0.3,
                        hashable >= 1.2 && < 1.4,
-                       lens >=4.4 && <5,
                        newtype >=0.2 && <0.3,
                        semigroups >= 0.18 && < 0.20
